bug fixed

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required# ,fresh_jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -16,7 +16,7 @@ class Item(Resource):
                         help="Every item needs a store_id."
                         )
 
-    @jwt_required()  # No longer needs brackets
+    @jwt_required  # No longer needs brackets
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -38,7 +38,13 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        '''
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+        '''
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
